+++ packages/database-postgres/entity/Block.ts
@@ -80,7 +80,7 @@ export class Block {
 
   @Column({
     default: 0,
-    type: 'bigint',
+    type: 'integer',
     nullable: false,
   })
   public _version_: number;
