+++ packages/database-postgres/entity/Transfer.ts
@@ -63,7 +63,7 @@ export class Transfer {
 
   @Column({
     default: 0,
-    type: 'bigint',
+    type: 'integer',
     nullable: false,
   })
   public _version_: number;
